#20 - Merge abstract class with Migration Command
