+++ src/Console/Commands/Migration.php
@@ -14,18 +14,42 @@ declare(strict_types=1);
 namespace Phalcon\Migrations\Console\Commands;
 
 use Phalcon\Config;
+use Phalcon\Cop\Parser;
 use Phalcon\Migrations\Console\Color;
+use Phalcon\Migrations\Console\Path;
 use Phalcon\Migrations\Migrations;
 use Phalcon\Migrations\Script\ScriptException;
 use Phalcon\Mvc\Model\Exception;
+use Phalcon\Config\Adapter\Ini as IniConfig;
+use Phalcon\Config\Adapter\Json as JsonConfig;
+use Phalcon\Config\Adapter\Yaml as YamlConfig;
 
 /**
  * Migration Command
  *
  * Generates/Run a migration
  */
-class Migration extends Command
+class Migration implements CommandsInterface
 {
+    /**
+     * @var Parser
+     */
+    protected $parser;
+
+    /**
+     * @var Path
+     */
+    protected $path;
+
+    /**
+     * @param Parser $parser
+     */
+    final public function __construct(Parser $parser)
+    {
+        $this->parser = $parser;
+        $this->path = new Path();
+    }
+
     /**
      * @return array
      */
@@ -213,5 +237,100 @@ class Migration extends Command
         }
 
         return $tables;
+    }
+
+    /**
+     * @param string $path Config path
+     *
+     * @return Config
+     * @throws CommandsException
+     */
+    protected function getConfig(string $path): Config
+    {
+        foreach (['app/config/', 'config/'] as $configPath) {
+            foreach (['ini', 'php', 'json', 'yaml', 'yml'] as $extension) {
+                if (file_exists($path . $configPath . "config." . $extension)) {
+                    return $this->loadConfig($path . $configPath . "/config." . $extension);
+                }
+            }
+        }
+
+        $directory = new \RecursiveDirectoryIterator('.');
+        $iterator = new \RecursiveIteratorIterator($directory);
+        foreach ($iterator as $f) {
+            if (preg_match('/config\.(php|ini|json|yaml|yml)$/i', $f->getPathName())) {
+                return $this->loadConfig($f->getPathName());
+            }
+        }
+
+        throw new CommandsException("Builder can't locate the configuration file.");
+    }
+
+    /**
+     * Determines correct adapter by file name
+     * and load config
+     *
+     * @param string $fileName Config file name
+     *
+     * @return Config
+     * @throws CommandsException
+     */
+    protected function loadConfig(string $fileName): Config
+    {
+        $pathInfo = pathinfo($fileName);
+
+        if (!isset($pathInfo['extension'])) {
+            throw new CommandsException('Config file extension not found.');
+        }
+
+        $extension = strtolower(trim($pathInfo['extension']));
+
+        switch ($extension) {
+            case 'php':
+                $config = include($fileName);
+                if (is_array($config)) {
+                    $config = new Config($config);
+                }
+
+                return $config;
+
+            case 'ini':
+                return new IniConfig($fileName);
+
+            case 'json':
+                return new JsonConfig($fileName);
+
+            case 'yaml':
+            case 'yml':
+                return new YamlConfig($fileName);
+
+            default:
+                throw new CommandsException("Builder can't locate the configuration file.");
+        }
+    }
+
+    /**
+     * Prints the available options in the script
+     *
+     * @param array $parameters
+     */
+    protected function printParameters($parameters): void
+    {
+        $length = 0;
+        foreach ($parameters as $parameter => $description) {
+            if ($length == 0) {
+                $length = strlen($parameter);
+            }
+
+            if (strlen($parameter) > $length) {
+                $length = strlen($parameter);
+            }
+        }
+
+        print Color::head('Options:') . PHP_EOL;
+        foreach ($parameters as $parameter => $description) {
+            print Color::colorize(' --' . $parameter . str_repeat(' ', $length - strlen($parameter)), Color::FG_GREEN);
+            print Color::colorize('    ' . $description) . PHP_EOL;
+        }
     }
 }
